* Corrected IBU claulation * Rename file extension to grsx

+++ BSImporter/BSClasses.cs
@@ -3,7 +3,7 @@
     public class BSGrainBill
     {
         public string FermentableName { set; get; }
-        public double AmountPercent { set; get; }
+        public int AmountPercent { set; get; }
     }
 
     public class BSHops
@@ -11,7 +11,7 @@
         public string Name { get; set; }
         public double AlphaAcid { get; set; }
         public double DryHopDay { get; internal set; }
-        public double BoilTime { get; internal set; }
+        public int BoilTime { get; internal set; }
         public double Amount { get; internal set; }
     }
